// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CurvaDeLocadoraResponse: {
            locadora: string;
            pontos: components["schemas"]["PontoDaCurvaResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}